pending tuning

+++ config.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Config struct {
-	MaxWorkerNum  uint
-	MinWorkerNum  uint
+	MaxWorkerNum  uint64
+	MinWorkerNum  uint64
 	MaxWorkerLife time.Duration
 	Debug         bool
 }
